Move some functionality into a util module for reuse
